+++ html/elements/frame1.php
@@ -1,0 +1,20 @@
+<?php require_once __DIR__ . '/prepend.php'; ?>
+
+<?php 
+
+  define('PAGE_TITLE', 'Frame1');
+
+  renderCoreTemplateStart(); 
+ 
+?>
+
+  <div class="single-column">
+
+    <p>First frame.</p>
+
+    <p><img src="/assets/img/city11.png" alt="city11"/></p>
+
+  </div>
+
+  
+<?php renderCoreTemplateEnd(); ?>